--- management/server/activity/mock.go
@@ -1,1 +1,0 @@
-package activity
