+++ management/server/nameserver_test.go
@@ -1057,7 +1057,7 @@ func createNSManager(t *testing.T) (*DefaultAccountManager, error) {
 	if err != nil {
 		return nil, err
 	}
-	eventStore := &activity.NoopEventStore{}
+	eventStore := &activity.InMemoryEventStore{}
 	return BuildManager(store, NewPeersUpdateManager(), nil, "", "", eventStore)
 }
