+++ management/server/route_test.go
@@ -937,7 +937,7 @@ func createRouterManager(t *testing.T) (*DefaultAccountManager, error) {
 	if err != nil {
 		return nil, err
 	}
-	eventStore := &activity.NoopEventStore{}
+	eventStore := &activity.InMemoryEventStore{}
 	return BuildManager(store, NewPeersUpdateManager(), nil, "", "", eventStore)
 }
